Added alternative PivotUI2
PivotUI generates HTML using a table which has a fixed layout.  The alternative UI generates a set of <div> tags.  The presentation of these tags can be defined using only CSS. The updated ./dist/pivot.css includes CSS that create a facsimile of the table presentation.  The facsimile is not exactly the same but it can be with a little more effort.