Update pivot.css
Updated the CSS for the <div> layout to more closely resemble the table layout.  Also added directives to illustrate how the <div> layout can be responsive.  The directive change the layout to a stacked presentation when the display device is a handheld one such as a mobile phone or if the page width is 800px or less.